requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.4
pillow>=10.0.0
google-api-python-client>=2.151.0
google-auth>=2.35.0
